+++ src/utils.rs
@@ -1,9 +1,9 @@
-use std::ffi::CStr;
+use std::ffi::{CStr, CString};
 use std::os::raw::c_char;
 
 use libheif_sys as lh;
 
-use crate::FileTypeResult;
+use crate::{FileTypeResult, HeifError, HeifErrorCode, HeifErrorSubCode};
 
 #[inline]
 pub(crate) fn cstr_to_str<'a>(c_str: *const c_char) -> Option<&'a str> {
@@ -16,6 +16,14 @@ pub(crate) fn cstr_to_str<'a>(c_str: *const c_char) -> Option<&'a str> {
             Err(_) => None,
         }
     }
+}
+
+pub(crate) fn str_to_cstring(s: &str, name: &str) -> Result<CString, HeifError> {
+    CString::new(s).map_err(|e| HeifError {
+        code: HeifErrorCode::UsageError,
+        sub_code: HeifErrorSubCode::InvalidParameterValue,
+        message: format!("Invalid value of '{}': {}", name, e),
+    })
 }
 
 /// Check file type by it first bytes.